lns+ 1.0.0 integration

+++ lib_cxng/include/lcx_aes.h
@@ -1,7 +1,7 @@
 
 /*******************************************************************************
 *   Ledger Nano S - Secure firmware
-*   (c) 2021 Ledger
+*   (c) 2022 Ledger
 *
 *  Licensed under the Apache License, Version 2.0 (the "License");
 *  you may not use this file except in compliance with the License.
@@ -46,7 +46,7 @@
  *
  * @param[in]  rawkey  Pointer to the supplied key.
  * 
- * @param[in]  key_len Length of the key: 8, 16 or 24 octets.
+ * @param[in]  key_len Length of the key: 16, 24 or 32 octets.
  * 
  * @param[out] key     Pointer to the key.
  * 
@@ -65,7 +65,7 @@ cx_err_t cx_aes_init_key_no_throw(const uint8_t *rawkey, size_t key_len, cx_aes_
  *
  * @param[in]  rawkey  Pointer to the supplied key.
  * 
- * @param[in]  key_len Length of the key: 8, 16 or 24 octets.
+ * @param[in]  key_len Length of the key: 16, 24 or 32 octets.
  * 
  * @param[out] key     Pointer to the key.
  * 
@@ -73,7 +73,7 @@ cx_err_t cx_aes_init_key_no_throw(const uint8_t *rawkey, size_t key_len, cx_aes_
  * 
  * @throw              CX_INVALID_PARAMETER
  */
-static inline int cx_aes_init_key ( const unsigned char * rawkey, unsigned int key_len, cx_aes_key_t * key )
+static inline size_t cx_aes_init_key ( const unsigned char * rawkey, unsigned int key_len, cx_aes_key_t * key )
 {
   CX_THROW(cx_aes_init_key_no_throw(rawkey, key_len, key));
   return key_len;
@@ -97,6 +97,9 @@ static inline int cx_aes_init_key ( const unsigned char * rawkey, unsigned int k
  *                     - CX_CHAIN_ECB
  *                     - CX_CHAIN_CBC
  *                     - CX_CHAIN_CTR
+ * 
+ *                    When using the CTR mode with AES, CX_ENCRYPT must be used for encryption
+ *                    and decryption.
  * 
  * @param[in] iv      Initialization vector.
  * 
@@ -151,6 +154,9 @@ cx_err_t cx_aes_iv_no_throw(const cx_aes_key_t *key,
  *                     - CX_CHAIN_CBC
  *                     - CX_CHAIN_CTR
  * 
+ *                    When using the CTR mode with AES, CX_ENCRYPT must be used for encryption
+ *                    and decryption.
+ * 
  * @param[in] iv      Initialization vector.
  * 
  * @param[in] iv_len  Length of the initialization vector.
@@ -173,7 +179,7 @@ cx_err_t cx_aes_iv_no_throw(const cx_aes_key_t *key,
  * @throws            CX_INVALID_PARAMETER
  * @throws            INVALID_PARAMETER
  */
-static inline int cx_aes_iv ( const cx_aes_key_t * key, int mode, unsigned char * iv, unsigned int iv_len, const unsigned char * in, unsigned int in_len, unsigned char * out, unsigned int out_len )
+static inline size_t cx_aes_iv ( const cx_aes_key_t * key, uint32_t mode, unsigned char * iv, unsigned int iv_len, const unsigned char * in, unsigned int in_len, unsigned char * out, unsigned int out_len )
 {
   size_t out_len_ = out_len;
   CX_THROW(cx_aes_iv_no_throw(key, mode, iv, iv_len, in, in_len, out, &out_len_));
@@ -264,7 +270,7 @@ cx_err_t cx_aes_no_throw(const cx_aes_key_t *key, uint32_t mode, const uint8_t *
  * @throws            CX_INVALID_PARAMETER
  * @throws            INVALID_PARAMETER
  */
-static inline int cx_aes ( const cx_aes_key_t * key, int mode, const unsigned char * in, unsigned int in_len, unsigned char * out, unsigned int out_len )
+static inline size_t cx_aes ( const cx_aes_key_t * key, uint32_t mode, const unsigned char * in, unsigned int in_len, unsigned char * out, unsigned int out_len )
 {
   size_t out_len_ = out_len;
   CX_THROW(cx_aes_no_throw(key, mode, in, in_len, out, &out_len_));
